+++ appGUI/preferences/excellon/ExcellonOptPrefGroupUI.py
@@ -29,7 +29,7 @@ class ExcellonOptPrefGroupUI(OptionsGroupUI):
         self.decimals = decimals
 
         # ## Create CNC Job
-        self.cncjob_label = FCLabel('<b>%s</b>' % _('Create CNCJob'))
+        self.cncjob_label = FCLabel('<b>%s</b>' % _('Parameters'))
         self.cncjob_label.setToolTip(
             _("Parameters used to create a CNC Job object\n"
               "for this drill object.")
@@ -41,53 +41,6 @@ class ExcellonOptPrefGroupUI(OptionsGroupUI):
         grid2.setColumnStretch(0, 0)
         grid2.setColumnStretch(1, 1)
 
-        # Operation Type
-        self.operation_label = FCLabel('<b>%s:</b>' % _('Operation'))
-        self.operation_label.setToolTip(
-            _("Operation type:\n"
-              "- Drilling -> will drill the drills/slots associated with this tool\n"
-              "- Milling -> will mill the drills/slots")
-        )
-        self.operation_radio = RadioSet(
-            [
-                {'label': _('Drilling'), 'value': 'drill'},
-                {'label': _("Milling"), 'value': 'mill'}
-            ]
-        )
-
-        grid2.addWidget(self.operation_label, 0, 0)
-        grid2.addWidget(self.operation_radio, 0, 1)
-
-        self.mill_type_label = FCLabel('%s:' % _('Milling Type'))
-        self.mill_type_label.setToolTip(
-            _("Milling type:\n"
-              "- Drills -> will mill the drills associated with this tool\n"
-              "- Slots -> will mill the slots associated with this tool\n"
-              "- Both -> will mill both drills and mills or whatever is available")
-        )
-        self.milling_type_radio = RadioSet(
-            [
-                {'label': _('Drills'), 'value': 'drills'},
-                {'label': _("Slots"), 'value': 'slots'},
-                {'label': _("Both"), 'value': 'both'},
-            ]
-        )
-
-        grid2.addWidget(self.mill_type_label, 1, 0)
-        grid2.addWidget(self.milling_type_radio, 1, 1)
-
-        self.mill_dia_label = FCLabel('%s:' % _('Milling Diameter'))
-        self.mill_dia_label.setToolTip(
-            _("The diameter of the tool who will do the milling")
-        )
-
-        self.mill_dia_entry = FCDoubleSpinner()
-        self.mill_dia_entry.set_precision(self.decimals)
-        self.mill_dia_entry.set_range(0.0000, 10000.0000)
-
-        grid2.addWidget(self.mill_dia_label, 2, 0)
-        grid2.addWidget(self.mill_dia_entry, 2, 1)
-
         # ### Milling Holes ## ##
         self.mill_hole_label = FCLabel('<b>%s</b>' % _('Mill Holes'))
         self.mill_hole_label.setToolTip(
@@ -97,7 +50,8 @@ class ExcellonOptPrefGroupUI(OptionsGroupUI):
 
         tdlabel = FCLabel('%s:' % _('Drill Tool dia'))
         tdlabel.setToolTip(
-            _("Diameter of the cutting tool.")
+            _("Diameter of the cutting tool\n"
+              "when milling drill holes.")
         )
         self.tooldia_entry = FCDoubleSpinner()
         self.tooldia_entry.set_precision(self.decimals)
@@ -109,7 +63,7 @@ class ExcellonOptPrefGroupUI(OptionsGroupUI):
         stdlabel = FCLabel('%s:' % _('Slot Tool dia'))
         stdlabel.setToolTip(
             _("Diameter of the cutting tool\n"
-              "when milling slots.")
+              "when milling slot holes.")
         )
         self.slot_tooldia_entry = FCDoubleSpinner()
         self.slot_tooldia_entry.set_precision(self.decimals)